Make generic types in lieu of duplicating code for spec types with compile-time bounds
