+++ src/domains.rs
@@ -1,3 +1,6 @@
+use crate::primitives::{Domain, Root};
+use ssz_rs::prelude::*;
+
 pub enum DomainType {
     BeaconProposer,
     BeaconAttester,
@@ -6,4 +9,10 @@ pub enum DomainType {
     VoluntaryExit,
     SelectionProof,
     AggregateAndProof,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct SigningData {
+    pub object_root: Root,
+    pub domain: Domain,
 }
