+++ src/lib.rs
@@ -1,7 +1,5 @@
 pub mod altair;
-pub mod configs;
 pub mod crypto;
 pub mod domains;
 pub mod phase0;
-pub mod presets;
 pub mod primitives;
