+++ src/phase0/beacon_block.rs
@@ -1,8 +1,72 @@
-use crate::{
-    crypto::BLSSignature,
-    primitives::{Root, Slot, ValidatorIndex},
+use crate::crypto::BLSSignature;
+use crate::phase0::operations::{
+    Attestation, AttesterSlashing, Deposit, Eth1Data, ProposerSlashing, SignedVoluntaryExit,
 };
+use crate::primitives::{Bytes32, Slot, ValidatorIndex};
 use ssz_rs::prelude::*;
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct BeaconBlockBody<
+    const MAX_PROPOSER_SLASHINGS: usize,
+    const MAX_VALIDATORS_PER_COMMITTEE: usize,
+    const MAX_ATTESTER_SLASHINGS: usize,
+    const MAX_ATTESTATIONS: usize,
+    const MAX_DEPOSITS: usize,
+    const MAX_VOLUNTARY_EXITS: usize,
+> {
+    pub randao_reveal: BLSSignature,
+    pub eth1_data: Eth1Data,
+    pub graffiti: Bytes32,
+    pub proposer_slashings: List<ProposerSlashing, MAX_PROPOSER_SLASHINGS>,
+    pub attester_slashings:
+        List<AttesterSlashing<MAX_VALIDATORS_PER_COMMITTEE>, MAX_ATTESTER_SLASHINGS>,
+    pub attestations: List<Attestation<MAX_VALIDATORS_PER_COMMITTEE>, MAX_ATTESTATIONS>,
+    pub deposits: List<Deposit, MAX_DEPOSITS>,
+    pub voluntary_exits: List<SignedVoluntaryExit, MAX_VOLUNTARY_EXITS>,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct BeaconBlock<
+    const MAX_PROPOSER_SLASHINGS: usize,
+    const MAX_VALIDATORS_PER_COMMITTEE: usize,
+    const MAX_ATTESTER_SLASHINGS: usize,
+    const MAX_ATTESTATIONS: usize,
+    const MAX_DEPOSITS: usize,
+    const MAX_VOLUNTARY_EXITS: usize,
+> {
+    pub slot: Slot,
+    pub proposer_index: ValidatorIndex,
+    pub parent_root: Root,
+    pub state_root: Root,
+    pub body: BeaconBlockBody<
+        MAX_PROPOSER_SLASHINGS,
+        MAX_VALIDATORS_PER_COMMITTEE,
+        MAX_ATTESTER_SLASHINGS,
+        MAX_ATTESTATIONS,
+        MAX_DEPOSITS,
+        MAX_VOLUNTARY_EXITS,
+    >,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct SignedBeaconBlock<
+    const MAX_PROPOSER_SLASHINGS: usize,
+    const MAX_VALIDATORS_PER_COMMITTEE: usize,
+    const MAX_ATTESTER_SLASHINGS: usize,
+    const MAX_ATTESTATIONS: usize,
+    const MAX_DEPOSITS: usize,
+    const MAX_VOLUNTARY_EXITS: usize,
+> {
+    pub message: BeaconBlock<
+        MAX_PROPOSER_SLASHINGS,
+        MAX_VALIDATORS_PER_COMMITTEE,
+        MAX_ATTESTER_SLASHINGS,
+        MAX_ATTESTATIONS,
+        MAX_DEPOSITS,
+        MAX_VOLUNTARY_EXITS,
+    >,
+    pub signature: BLSSignature,
+}
 
 #[derive(Default, Debug, SimpleSerialize)]
 pub struct BeaconBlockHeader {
