+++ src/phase0/beacon_state.rs
@@ -1,0 +1,62 @@
+use crate::phase0::beacon_block::BeaconBlockHeader;
+use crate::phase0::fork::Fork;
+use crate::phase0::operations::{Checkpoint, Eth1Data, PendingAttestation};
+use crate::phase0::validator::Validator;
+use crate::primitives::{Bytes32, Gwei, Root, Slot, JUSTIFICATION_BITS_LENGTH};
+use ssz_rs::prelude::*;
+
+pub(super) const fn get_eth1_data_votes_bound(
+    epochs_per_eth1_voting_period: usize,
+    slots_per_epoch: usize,
+) -> usize {
+    epochs_per_eth1_voting_period * slots_per_epoch
+}
+
+pub(super) const fn get_pending_attestations_bound(
+    max_attestations: usize,
+    slots_per_epoch: usize,
+) -> usize {
+    max_attestations * slots_per_epoch
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct HistoricalBatch<const SLOTS_PER_HISTORICAL_ROOT: usize> {
+    pub block_roots: Vector<Root, SLOTS_PER_HISTORICAL_ROOT>,
+    pub state_roots: Vector<Root, SLOTS_PER_HISTORICAL_ROOT>,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct BeaconState<
+    const SLOTS_PER_HISTORICAL_ROOT: usize,
+    const HISTORICAL_ROOTS_LIMIT: usize,
+    const ETH1_DATA_VOTES_BOUND: usize,
+    const VALIDATOR_REGISTRY_LIMIT: usize,
+    const EPOCHS_PER_HISTORICAL_VECTOR: usize,
+    const EPOCHS_PER_SLASHINGS_VECTOR: usize,
+    const MAX_VALIDATORS_PER_COMMITTEE: usize,
+    const PENDING_ATTESTATIONS_BOUND: usize,
+> {
+    pub genesis_time: u64,
+    pub genesis_validators_root: Root,
+    pub slot: Slot,
+    pub fork: Fork,
+    pub latest_block_header: BeaconBlockHeader,
+    pub block_roots: Vector<Root, SLOTS_PER_HISTORICAL_ROOT>,
+    pub state_roots: Vector<Root, SLOTS_PER_HISTORICAL_ROOT>,
+    pub historical_roots: List<Root, HISTORICAL_ROOTS_LIMIT>,
+    pub eth1_data: Eth1Data,
+    pub eth1_data_votes: List<Eth1Data, ETH1_DATA_VOTES_BOUND>,
+    pub eth1_deposit_index: u64,
+    pub validators: List<Validator, VALIDATOR_REGISTRY_LIMIT>,
+    pub balances: List<Gwei, VALIDATOR_REGISTRY_LIMIT>,
+    pub randao_mixes: Vector<Bytes32, EPOCHS_PER_HISTORICAL_VECTOR>,
+    pub slashings: Vector<Gwei, EPOCHS_PER_SLASHINGS_VECTOR>,
+    pub previous_epoch_attestations:
+        List<PendingAttestation<MAX_VALIDATORS_PER_COMMITTEE>, PENDING_ATTESTATIONS_BOUND>,
+    pub current_epoch_attestations:
+        List<PendingAttestation<MAX_VALIDATORS_PER_COMMITTEE>, PENDING_ATTESTATIONS_BOUND>,
+    pub justification_bits: Bitvector<JUSTIFICATION_BITS_LENGTH>,
+    pub previous_justified_checkpoint: Checkpoint,
+    pub current_justified_checkpoint: Checkpoint,
+    pub finalized_checkpoint: Checkpoint,
+}
