+++ src/phase0/configs/mod.rs
@@ -1,0 +1,2 @@
+pub mod mainnet;
+pub mod minimal;
