+++ src/phase0/operations.rs
@@ -1,13 +1,80 @@
-use crate::crypto::BLSSignature;
+use crate::crypto::{BLSPubkey, BLSSignature};
 use crate::phase0::beacon_block::SignedBeaconBlockHeader;
-use crate::phase0::DepositData;
-use crate::primitives::{Bytes32, Epoch, ValidatorIndex, DEPOSIT_CONTRACT_TREE_DEPTH};
+use crate::primitives::{
+    Bytes32, CommitteeIndex, Epoch, Gwei, Hash32, Root, Slot, ValidatorIndex,
+    DEPOSIT_CONTRACT_TREE_DEPTH,
+};
 use ssz_rs::prelude::*;
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct Checkpoint {
+    pub epoch: Epoch,
+    pub root: Root,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct AttestationData {
+    pub slot: Slot,
+    pub index: CommitteeIndex,
+    pub beacon_block_root: Root,
+    pub source: Checkpoint,
+    pub target: Checkpoint,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct IndexedAttestation<const MAX_VALIDATORS_PER_COMMITTEE: usize> {
+    pub attesting_indices: List<ValidatorIndex, MAX_VALIDATORS_PER_COMMITTEE>,
+    pub data: AttestationData,
+    pub signature: BLSSignature,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct PendingAttestation<const MAX_VALIDATORS_PER_COMMITTEE: usize> {
+    pub aggregation_bits: Bitlist<MAX_VALIDATORS_PER_COMMITTEE>,
+    pub data: AttestationData,
+    pub inclusion_delay: Slot,
+    pub proposer_index: ValidatorIndex,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct Attestation<const MAX_VALIDATORS_PER_COMMITTEE: usize> {
+    pub aggregation_bits: Bitlist<MAX_VALIDATORS_PER_COMMITTEE>,
+    pub data: AttestationData,
+    pub signature: BLSSignature,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct Eth1Data {
+    pub deposit_root: Root,
+    pub deposit_count: u64,
+    pub block_hash: Hash32,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct DepositMessage {
+    pub pubkey: BLSPubkey,
+    pub withdrawal_credentials: Bytes32,
+    pub amount: Gwei,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct DepositData {
+    pub pubkey: BLSPubkey,
+    pub withdrawal_credentials: Bytes32,
+    pub amount: Gwei,
+    pub signature: BLSSignature,
+}
 
 #[derive(Default, Debug, SimpleSerialize)]
 pub struct ProposerSlashing {
     pub signed_header_1: SignedBeaconBlockHeader,
     pub signed_header_2: SignedBeaconBlockHeader,
+}
+
+#[derive(Default, Debug, SimpleSerialize)]
+pub struct AttesterSlashing<const MAX_VALIDATORS_PER_COMMITTEE: usize> {
+    pub attestation_1: IndexedAttestation<MAX_VALIDATORS_PER_COMMITTEE>,
+    pub attestation_2: IndexedAttestation<MAX_VALIDATORS_PER_COMMITTEE>,
 }
 
 const fn get_deposit_proof_length() -> usize {
